Use resource instead of base
