+++ lib/ncs_analytics.rb
@@ -1,7 +1,7 @@
 require 'ncs_analytics/version'
 require 'ncs_analytics/configuration'
 require 'ncs_analytics/errors'
-require 'ncs_analytics/base'
+require 'ncs_analytics/resource'
 require 'ncs_analytics/client'
 
 module NcsAnalytics
