+++ lib/ncs_analytics/resource/plant_batch.rb
@@ -1,5 +1,13 @@
 module NcsAnalytics
-  class Harvest < NcsAnalytics::Base
+  class PlantBatch < NcsAnalytics::Resource
+    RESOURCE_NAME = "#{name.split('::').last.downcase}es".to_sym
+
+    def find(id = nil)
+      raise Errors::BadRequest, 'A plant ID is required' unless id
+
+      get("plantid/#{id}")
+    end
+
     def active
       by_state(:active)
     end
@@ -8,44 +16,28 @@ module NcsAnalytics
       by_state(:inactive)
     end
 
-    def onhold
-      by_state(:onhold)
-    end
-
     def create(payload)
       # TODO: Validate the payload
       payload = [payload] unless payload.is_a?(Array)
-      post(:create, payload)
+      post(:createplantings, payload)
     end
 
-    def create_package(payload)
+    def create_packages(payload)
       # TODO: Validate the payload
       payload = [payload] unless payload.is_a?(Array)
       post(:createpackages, payload)
     end
 
-    def update(payload)
+    def destroy(payload)
       # TODO: Validate the payload
       payload = [payload] unless payload.is_a?(Array)
-      post(:update, payload)
+      post(:destroy, payload)
     end
 
-    def finish(payload)
+    def change_growth_phase(payload)
       # TODO: Validate the payload
       payload = [payload] unless payload.is_a?(Array)
-      post(:finish, payload)
-    end
-
-    def unfinish(payload)
-      # TODO: Validate the payload
-      payload = [payload] unless payload.is_a?(Array)
-      post(:unfinish, payload)
-    end
-
-    def remove_waste(payload)
-      # TODO: Validate the payload
-      payload = [payload] unless payload.is_a?(Array)
-      post(:removewaste, payload)
+      post(:changegrowthphase, payload)
     end
 
     private
