+++ spec/ncs_analytics/resource/harvest_spec.rb
