+++ spec/ncs_analytics/resource/item_spec.rb
