+++ spec/ncs_analytics/resource/package_spec.rb
