+++ spec/ncs_analytics/resource/plant_batch_spec.rb
