+++ spec/ncs_analytics/resource/plant_spec.rb
