+++ spec/ncs_analytics/resource/room_spec.rb
