+++ spec/ncs_analytics/resource/sale_spec.rb
