+++ spec/ncs_analytics/resource/strain_spec.rb
