+++ spec/ncs_analytics/resource/test_spec.rb
